Move stylesheet object to util file
